HSEARCH-3834 Extract utils allowing to inspect the index content from LuceneIndexWriterCommitIT
